test: extract fullnode-specific internal tests
Move debug methods and storage reload tests into
fullnode-specific/internal.test.ts. Removes the entire
'internal methods' describe block from the monolithic
hathorwallet_others.test.ts file.
